// src/app/models/newOrder.ts
export class NewOrder{

        custID: number;
        empid: number;
        orderDate: string
        requiredDate: string;
        shippeddate: string;
        shipperId: number;
        freigth: number;
        shipname: string;
        shipaddress: string;
        shipcity: string;
        shiregion: string;
        shipppostalcode: string;
        shipcountry: string;
        productId: number;
        unitPrice: number;
        qty: number;
        discount: number
}